pyproject.toml
[project]
name = "familydashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]
